Move 'rpc::protocol::Receiver' into 'rpc::transaction'
Since the 'Receiver' struct is more relevant to the _building_ of the
unsigned protocol txs than to the multisig crypto (or any other protocol
concern), it probably belongs in the 'transaction' module with the rest
of the tx builder logic.
